+++ baybe/objective.py
@@ -10,9 +10,10 @@ import pandas as pd
 from attr import define, field
 from attr.validators import deep_iterable, in_, instance_of, min_len
 
+from baybe.serialization import SerialMixin
 from baybe.targets.base import Target
 from baybe.targets.numerical import NumericalTarget
-from baybe.utils import SerialMixin, geom_mean
+from baybe.utils import geom_mean
 
 
 def _normalize_weights(weights: List[float]) -> List[float]: