+++ baybe/serialization/__init__.py
@@ -1,0 +1,24 @@
+"""Serialization functionality."""
+
+from baybe.serialization.core import (
+    block_deserialization_hook,
+    block_serialization_hook,
+    converter,
+    get_base_structure_hook,
+    select_constructor_hook,
+    unstructure_base,
+)
+from baybe.serialization.mixin import SerialMixin
+from baybe.serialization.utils import deserialize_dataframe, serialize_dataframe
+
+__all__ = [
+    "block_deserialization_hook",
+    "block_serialization_hook",
+    "converter",
+    "deserialize_dataframe",
+    "get_base_structure_hook",
+    "select_constructor_hook",
+    "serialize_dataframe",
+    "unstructure_base",
+    "SerialMixin",
+]