+++ baybe/surrogates/base.py
@@ -10,9 +10,9 @@ from attr import define, field
 from torch import Tensor
 
 from baybe.searchspace import SearchSpace
+from baybe.serialization import SerialMixin, converter, unstructure_base
 from baybe.surrogates.utils import _prepare_inputs, _prepare_targets
-from baybe.utils import SerialMixin
-from baybe.utils.serialization import converter, get_subclasses, unstructure_base
+from baybe.utils import get_subclasses
 
 # Define constants
 _MIN_VARIANCE = 1e-6